Separated the different scripts into three folders. One for all the different game objects, one for utility functions that are used a lot, and another for shared objects and one time use functions
