+++ public/index.html
@@ -50,7 +50,7 @@
     <div class="centered">
         <canvas id="canvas"></canvas>
     </div>
-    <script type="module" src="./dist/main.js"></script>
+    <script type="module" src="./dist/index.js"></script>
 </body>
 
 </html>
