+++ src/game/OnetimeOrShared/resolution.ts
