+++ src/game/README.md
@@ -1,0 +1,1 @@
+### File layout
