+++ src/index.ts
@@ -1,1 +1,2 @@
-console.log("Hello World!");
+import Game from './game'
+new Game({});
