--- src/main.ts
@@ -1,2 +1,0 @@
-import Game from './game'
-new Game({});
